fixing linter errors on accounts new
